Add VS 2022 version; ignore C4127, pthread_cleanup_pop(1 /* always call cleanup handler */); if (1 == 1) { } (X -> √); fix warning C4702 Inaccessible code; fix warning C4310 Cast truncated constant value;

+++ tests/cancel4.c
@@ -99,7 +99,7 @@ static bag_t threadbag[NUMTHREADS + 1];
 static void *
 mythread(void * arg)
 {
-  void* result = (void*)((int)(size_t)PTHREAD_CANCELED + 1);
+  void* result = (void*)((intptr_t)PTHREAD_CANCELED + 1);
   bag_t * bag = (bag_t *) arg;
 
   assert(bag == &threadbag[bag->threadnum]);
